+++ test/bones_api_test.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.2.13
+// BUILDER: reflection_factory/1.2.14
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -36,7 +36,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule> {
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.13');
+  Version get reflectionFactoryVersion => Version.parse('1.2.14');
 
   @override
   MyInfoModule$reflection withObject([MyInfoModule? obj]) =>
@@ -594,7 +594,10 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule> {
               ParameterReflection(TypeReflection<APIRequest>(APIRequest),
                   'request', false, true, null, null)
             ],
-            null,
+            const <ParameterReflection>[
+              ParameterReflection(
+                  TypeReflection.tString, 'routeName', true, false, null, null)
+            ],
             null,
             null);
       case 'resolveroute':